requirements.txt
selenium==4.31.0
python-telegram-bot==20.7
